Day 5/rag-assistant/requirements.txt
fastapi
uvicorn[standard]
streamlit
python-dotenv
python-multipart
numpy
faiss-cpu
sentence-transformers
pymupdf
requests
pydantic

